Finished implementing ProductRepository

+++ FlameAndWax.Data/Models/ProductModel.cs
@@ -5,7 +5,7 @@
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public int ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
         public int QuantityPerUnit { get; set; }
         public double UnitPrice { get; set; }
         public int UnitsInStock { get; set; }
